Add files via upload
fundholdings1.html 
- is an xml to datatable parser. It will fetch data from the www.sec.gov website and display the data in table format which can then be exported to csv or xls format.

Requirements: 
- This file only works with Chrome (Version 92 or greater) and a CORS Extension that is allowed access to https://www.sec.gov/* An example of a CORS extension is "Moesif Origin & CORS Changer" which can be set to access specific sites.

Checksum:
SHA 256. - shasum -a 256 fundholdings1.html 
b569f96758b1672697e8915f5aa8f29d3380a475026cc79c2457d041b0bc98e8  


Disclaimer: Your use and access of the content in this repository are AT YOUR OWN RISK. You understand and agree that the services and content are provided to you on an "AS IS" and "AS AVAILABLE" basis. This repository and its publisher disclaim all warranties and conditions, whether express or implied, of merchantability, fitness for a particular purpose, or non-infrintement.

This repository and its publisher make no warranty or representation and disclaim all responsibility and liability for: (i) the completeness, accuracy, availability, timeliness, security or reliability of any code and content; (ii) any harm to your computer system, loss of data, or other harm that results from your access to or use of any content.

This repository and its publisher give no warranty and accepts no responsibility or liability for accuracy of the content and information. Under no circumstances will this repository or the publisher will be held responsible or liable in any way for any claims, damages, losses, expenses, costs or liabilities whatsoever (including without limitation, any direct or indirect damages for loss of profits, business interruption or loss of information) resulting or arising directly or indirectly from your use of or inability to use this content or any sites linked to it, or from your reliance on the information and content in this repository.